Move task list into Task component folder
